requirements.txt
attrs==23.2.0
blinker==1.9.0
charset-normalizer==3.3.2
click==8.3.1
cloud-init==24.4
configobj==5.0.8
dbus-python==1.3.2
decorator==5.1.1
distro==1.9.0
dnf==4.20.0
file-magic==0.4.0
Flask==3.1.2
idna==3.7
iniparse==0.5
itsdangerous==2.2.0
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==2.3
jsonschema==4.19.1
jsonschema-specifications==2023.11.2
libcomps==0.1.21
libdnf==0.73.1
MarkupSafe==2.1.3
mysql-connector-python==9.5.0
oauthlib==3.2.2
perf==0.1
PyGObject==3.46.0
pyinotify==0.9.6
PyMySQL==1.1.2
pyserial==3.5
python-dateutil==2.9.0.post0
python-linux-procfs==0.7.3
pyudev==0.24.1
PyYAML==6.0.1
pyynl @ file:///builddir/build/BUILD/kernel-6.12.0-55.34.1.el10_0/linux-6.12.0-55.34.1.el10_0.x86_64/tools/net/ynl
referencing==0.31.1
requests==2.32.4
rpds-py==0.17.1
rpm==4.19.1.1
selinux @ file:///builddir/build/BUILD/libselinux-3.8/src
sepolicy @ file:///builddir/build/BUILD/selinux-3.8/python/sepolicy
setools==4.5.1
setuptools==69.0.3
six==1.16.0
subscription-manager==1.30.6.1
systemd-python==235
urllib3==1.26.19
Werkzeug==3.1.4
